First working implementation of HTTP fallback support.
Add Http ConnectionAdapter that mimicks the WebSocket interface.
A client_id is now included with each event sent to a client. New
server events are streamed to HTTP clients via a persistent
connection. Consumers of the HTTP stream can send new events to
the server by POSTing back to the /websocket URL and referencing
their client_id in the request. The server finds the correct
connection in the open connecitons array and dispatches the new
event accordingly.