Major refactor to prepare for namespacing events and channels.
Reorganize classes to better separate their responsibilities. Add
an Event object to encapsulate event related data and make it easier
to pass events between methods. Rename the Events module to
EventMap to be a little more descriptive. Whiddle down the
responsibilities of the ConnectionManager to just the opening and
closing of connections.