Use the ruby redis driver instead of synchrony during shutdown.
This spawns a new Redis connection using the default ruby Redis driver
during shutdown to remove the server token from the active server tokens
set. It also creates this new connection within a new Thread.

This is slightly less efficient but solves the ruby 2.0 problem of
using a Mutex in a signal trap context. Closes #90.